fix: Add missing int32 cast in folderShareUseCase
- Fixed GetFolderFileCount call to cast share.Folder.UserID to int32

+++ services/cloudRepositoryService/features/cloudRepository/usecase/folderShareUseCase.go
@@ -180,7 +180,7 @@ func (u *FolderShareUseCase) GetSharedWithMeFolders(
 	for _, share := range shares {
 		if share.Folder != nil {
 			// Get file count for the folder
-			fileCount, err := u.FolderRepo.GetFolderFileCount(ctx, share.Folder.ID, share.Folder.UserID)
+			fileCount, err := u.FolderRepo.GetFolderFileCount(ctx, share.Folder.ID, int32(share.Folder.UserID))
 			if err != nil {
 				fileCount = 0 // Default to 0 on error
 			}
